requirements.txt
# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data Processing
pandas==2.1.4
numpy==1.26.2

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Dashboard
streamlit==1.30.0
plotly==5.18.0

# ML
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
